--- a/ansible/ansible.cfg
+++ b/ansible/ansible.cfg
@@ -1,4 +1,7 @@
 [defaults]
-deprecation_warnings=False
-inventory=/opt/retronas/ansible/hosts.yml
+deprecation_warnings = False
+command_warnings = False
+inventory = /opt/retronas/ansible/hosts.yml
 log_path = /opt/retronas/log/ansible.log
+callbacks_enabled = profile_tasks
+callback_whitelist = profile_tasks