--- a/ansible/ansible.cfg
+++ b/ansible/ansible.cfg
@@ -5,3 +5,4 @@ inventory = /opt/retronas/ansible/hosts.yml
 log_path = /opt/retronas/log/ansible.log
 callbacks_enabled = profile_tasks
 callback_whitelist = profile_tasks
+cache = True