--- a/ansible/hosts.yml
+++ b/ansible/hosts.yml
@@ -1,2 +1,3 @@
 [all]
 localhost ansible_connection=local
+cache_plugin=jsonfile