--- a/ansible/install_aria2.yml
+++ b/ansible/install_aria2.yml
@@ -1,4 +1,5 @@
 - hosts: localhost
+  gather_facts: no
 
   vars:
     - my_name: "aria2"