--- a/ansible/install_batocera_cifs.yml
+++ b/ansible/install_batocera_cifs.yml
@@ -45,7 +45,7 @@
     with_items:
       "{{ top_level_paths }}"
     when: 
-      - item.enabled == "yes"
+      - item.enabled is true
     notify: "{{ my_name }} - Restart service"
 
   - name: "{{ my_name }} - build layout list (set fact)"