--- a/ansible/install_cockpit.yml
+++ b/ansible/install_cockpit.yml
@@ -2,10 +2,16 @@
 - import_playbook: install_filesystems.yml
 
 - hosts: localhost
+  gather_facts: yes
 
   vars:
     - my_name: "Cockpit"
     - my_file: "install_cockpit"
+    - my_service: "cockpit"
+
+    - my_packages:
+      - gnupg
+      - aria2
 
   tasks:
 
@@ -13,12 +19,9 @@
     include_vars: retronas_vars.yml
     
   - name: "{{ my_name }} - install prerequisite packages"
-    apt:
-       name: "{{ item }}"
+    package:
+       name: "{{ my_packages }}"
        state: latest
-    with_items:
-      - gnupg
-      - aria2
 
 #  - name: "{{ my_name }} - Install Debian Signing Keys 1"
 #    shell: "/usr/bin/apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 648ACFD622F3D138"
@@ -67,6 +70,8 @@
   - name: "{{ my_name }} - Force update repo cache"
     shell:
       cmd: "/usr/bin/apt update || exit 0"
+    when: ansible_distribution == "Ubuntu" or
+          ansible_distribution == "Debian"
 
   - name: "{{ my_name }} - Install from repo"
     apt:
@@ -81,22 +86,18 @@
 
   - name: "{{ my_name }} - enable startup services"
     service:
-      name: "{{ item }}"
+      name: "{{ my_service }}"
       state: started
       enabled: yes
       daemon_reload: yes
-    with_items:
-      - cockpit
 
   handlers:
 
   - name: "{{ my_name }} - Restart service"
     service:
-      name: "{{ item }}"
+      name: "{{ my_service }}"
       state: restarted
       daemon_reload: yes
-    with_items:
-      - cockpit
 
 - import_playbook: retronas_system_config.yml
   vars: