--- a/ansible/install_emuelec_cifs.yml
+++ b/ansible/install_emuelec_cifs.yml
@@ -3,6 +3,7 @@
 - import_playbook: install_romdir.yml
 
 - hosts: localhost
+  gather_facts: no
 
   vars:
     - my_name: "EmuElec CIFS"