--- a/ansible/install_etherdfs.yml
+++ b/ansible/install_etherdfs.yml
@@ -1,19 +1,11 @@
 - hosts: localhost
+  gather_facts: no
 
   vars:
     - my_name: "EtherDFS"
     - my_file: "install_etherdfs"
 
-  tasks:
-
-  - name: "{{ my_name }} - Load RetroNAS config"
-    include_vars: retronas_vars.yml
-
-  - name: "{{ my_name }} - Install build tools"
-    apt:
-       name: "{{ item }}"
-       state: latest
-    with_items:
+    - my_packages:
       - make
       - automake
       - autoconf
@@ -22,6 +14,16 @@
       - git
       - build-essential
       - coreutils
+
+  tasks:
+
+  - name: "{{ my_name }} - Load RetroNAS config"
+    include_vars: retronas_vars.yml
+
+  - name: "{{ my_name }} - Install build tools"
+    package:
+       name: "{{ my_packages }}"
+       state: latest
 
   - name: "{{ my_name }} - create MS-DOS directory"
     file: