--- a/ansible/install_extract-xiso.yml
+++ b/ansible/install_extract-xiso.yml
@@ -1,8 +1,15 @@
 - hosts: localhost
+  gather_facts: no
 
   vars:
     - my_name: "extract-xiso"
     - my_file: "install_extract-xiso"
+
+    - my_packages: 
+      - make
+      - gcc
+      - coreutils
+      - cmake
 
   tasks:
 
@@ -10,14 +17,9 @@
     include_vars: retronas_vars.yml
 
   - name: "{{ my_name }} - Install build tools"
-    apt:
-       name: "{{ item }}"
+    package:
+       name: "{{ my_packages }}"
        state: latest
-    with_items:
-      - make
-      - gcc
-      - coreutils
-      - cmake
 
   - name: "{{ my_name }} - Generate scripts"
     template: