--- a/ansible/install_extradirs.yml
+++ b/ansible/install_extradirs.yml
@@ -5,10 +5,10 @@
     - my_file: "install_romdir"
 
     - top_level_paths:
-      - { name: "saves", enabled: "yes", systems: "yes" }
-      - { name: "savestates", enabled: "yes", systems: "yes" }
-      - { name: "bios", enabled: "yes", systems: "yes" }
-      - { name: "wallpapers", enabled: "yes", systems: "no" }
+      - { name: "saves", enabled: yes, systems: yes }
+      - { name: "savestates", enabled: yes, systems: yes }
+      - { name: "bios", enabled: yes, systems: yes }
+      - { name: "wallpapers", enabled: yes, systems: no }
 
   tasks:
     - name: "{{ my_name }}"