--- a/ansible/install_extradirs.yml
+++ b/ansible/install_extradirs.yml
@@ -5,9 +5,9 @@
     - my_file: "install_romdir"
 
     - top_level_paths:
-      - { name: "saves", enabled: yes, systems: yes }
+      - { name: "saves",      enabled: yes, systems: yes }
       - { name: "savestates", enabled: yes, systems: yes }
-      - { name: "bios", enabled: yes, systems: yes }
+      - { name: "bios",       enabled: yes, systems: yes }
       - { name: "wallpapers", enabled: yes, systems: no }
 
   tasks: