--- a/ansible/install_extradirs.yml
+++ b/ansible/install_extradirs.yml
@@ -1,4 +1,5 @@
 - hosts: localhost
+  gather_facts: no
 
   vars:
     - my_name: "Generic Extra dir(s)"