--- a/ansible/install_netatalk2.yml
+++ b/ansible/install_netatalk2.yml
@@ -1,4 +1,5 @@
 - hosts: localhost
+  gather_facts: yes
 
   vars:
     - my_name: "Netatalk2"