… ansible/install_fenrir-ode-webserver.yml
… ansible/install_filesystems.yml
… ansible/install_fsp.yml
… ansible/install_gogrepo.yml
… ansible/install_hb-store-cdn.yml
… ansible/install_hdldump.yml
… ansible/install_hdparm.yml
… ansible/install_lighttpd.yml
… ansible/install_litch.yml
… ansible/install_megatools.yml
… ansible/install_mister_cifs.yml
… ansible/install_nbd-client.yml
… ansible/install_netatalk2x.yml
… ansible/install_netatalk3.yml
… ansible/install_nfs.yml
… ansible/install_openssh.yml
… ansible/install_pfsshell.yml
… ansible/install_proftpd.yml
… ansible/install_ps2_openps2loader.yml
… ansible/install_ps3netsrv.yml
… ansible/install_pygopherd.yml
… ansible/install_recalbox_cifs.yml
… ansible/install_romdir.yml
… ansible/install_romimport.yml
… ansible/install_sabretools.yml
… ansible/install_samba.yml
… ansible/install_tcpser.yml
… ansible/install_telnet.yml
… ansible/install_tftpd-hpa.yml
… ansible/install_tnfs.yml