Speed hacks and general play optimisations, moving away from the defaulting ansibleness of big loops for directory creation and link creation which results is alot of performance gains at the cost of some fidelity

RPI4 - main
```
Saturday 19 November 2022  03:07:05 +0000 (0:00:00.701)       0:21:44.845 ***** 
=============================================================================== 
Generic Extra dir(s) - build layout -------------------------------------- 527.38s
MiSTer CIFS - build systems directory layout ----------------------------- 373.00s
get facts for path ------------------------------------------------------- 158.25s
Generic ROM dir - build layout ------------------------------------------- 154.85s
Generic Extra dir(s) - build x-system symlinks ---------------------------- 18.31s
MiSTer CIFS - removing old symlink layout see issue #21 -------------------- 6.27s
```

RPI4 - diropt
```
Saturday 19 November 2022  02:13:35 +0000 (0:00:00.719)       0:03:16.253 ***** 
=============================================================================== 
MiSTer CIFS - build layout list (set fact) -------------------------------------- 60.36s
Generic Extra dir(s) - build layout list (set fact) ----------------------------- 45.68s
Generic Extra dir(s) - build x-system symlinks ---------------------------------- 14.73s
```